+++ src/cache-save.ts
@@ -8,6 +8,7 @@ export async function run(): Promise<void> {
     await cacheRTXTools()
   } catch (error) {
     if (error instanceof Error) core.setFailed(error.message)
+    else throw error
   }
 }
