--- src/main.ts
@@ -1,127 +1,0 @@
-import * as cache from '@actions/cache'
-import * as core from '@actions/core'
-import * as exec from '@actions/exec'
-import * as glob from '@actions/glob'
-import * as fs from 'fs'
-import * as os from 'os'
-import * as path from 'path'
-import { rtxDir } from './utils'
-
-async function run(): Promise<void> {
-  await setToolVersions()
-  await setRtxToml()
-
-  if (core.getBooleanInput('cache')) {
-    await restoreRTXCache()
-    core.saveState('CACHE', false)
-  } else {
-    core.saveState('CACHE', true)
-    core.setOutput('cache-hit', false)
-  }
-
-  const version = core.getInput('version')
-  await setupRTX(version)
-  await setEnvVars()
-  await exec.exec('rtx', ['--version'])
-  const install = core.getBooleanInput('install', { required: false })
-  if (install) {
-    await exec.exec('rtx', ['install'])
-  }
-  await setPaths()
-}
-
-async function setEnvVars(): Promise<void> {
-  if (!process.env['RTX_TRUSTED_CONFIG_PATHS']) {
-    core.exportVariable(
-      'RTX_TRUSTED_CONFIG_PATHS',
-      path.join(process.cwd(), '.rtx.toml')
-    )
-  }
-  if (!process.env['RTX_YES']) {
-    core.exportVariable('RTX_YES', '1')
-  }
-}
-
-async function restoreRTXCache(): Promise<void> {
-  const cachePath = rtxDir()
-  const fileHash = await glob.hashFiles(`**/.tool-versions\n**/.rtx.toml`)
-  const primaryKey = `rtx-tools-${getOS()}-${os.arch()}-${fileHash}`
-
-  core.saveState('PRIMARY_KEY', primaryKey)
-
-  const cacheKey = await cache.restoreCache([cachePath], primaryKey)
-  core.setOutput('cache-hit', Boolean(cacheKey))
-
-  if (!cacheKey) {
-    core.info(`rtx cache not found for ${getOS()}-${os.arch()} tool versions`)
-    return
-  }
-
-  core.saveState('CACHE_KEY', cacheKey)
-  core.info(`rtx cache restored from key: ${cacheKey}`)
-}
-
-async function setupRTX(version: string | undefined): Promise<void> {
-  const rtxBinDir = path.join(rtxDir(), 'bin')
-  const url = version
-    ? `https://rtx.jdx.dev/v${version}/rtx-v${version}-${getOS()}-${os.arch()}`
-    : `https://rtx.jdx.dev/rtx-latest-${getOS()}-${os.arch()}`
-  await fs.promises.mkdir(rtxBinDir, { recursive: true })
-  await exec.exec('curl', [url, '--output', path.join(rtxBinDir, 'rtx')])
-  await exec.exec('chmod', ['+x', path.join(rtxBinDir, 'rtx')])
-  core.addPath(rtxBinDir)
-}
-
-// returns true if tool_versions was set
-async function setToolVersions(): Promise<boolean> {
-  const toolVersions = core.getInput('tool_versions', { required: false })
-  if (toolVersions) {
-    await writeFile('.tool-versions', toolVersions)
-    return true
-  }
-  return false
-}
-
-async function setRtxToml(): Promise<boolean> {
-  const toml = core.getInput('rtx_toml', { required: false })
-  if (toml) {
-    await writeFile('.rtx.toml', toml)
-    return true
-  }
-  return false
-}
-
-function getOS(): string {
-  switch (process.platform) {
-    case 'darwin':
-      return 'macos'
-    default:
-      return process.platform
-  }
-}
-
-async function setPaths(): Promise<void> {
-  for (const binPath of await getBinPaths()) {
-    core.addPath(binPath)
-  }
-}
-
-async function getBinPaths(): Promise<string[]> {
-  const output = await exec.getExecOutput('rtx', ['bin-paths'])
-  return output.stdout.split('\n')
-}
-
-const writeFile = async (p: fs.PathLike, body: string): Promise<void> =>
-  fs.promises.writeFile(p, body, { encoding: 'utf8' })
-
-if (require.main === module) {
-  try {
-    run()
-  } catch (err) {
-    if (err instanceof Error) {
-      core.setFailed(err.message)
-    } else throw err
-  }
-}
-
-export { run }
